convert template into JSX, and Combined loops to reduce redundancy.
